[tiger] Drop the one directory per chapter rule; all happens in tiger/
From now on, all development is done in the tiger directory. Each new chapter
is built on top of the previous chapter, instead of creating new directories
for each chapter.

I've kept the directories for the first two chapters because the first one
ins't actually part of the Tiger compiler, while the second one is not yet
finished.

Git branches bearing the chapter number and name will be used as milestones
for the chapters I've finished.