[lingua-003] Add identifier references to the language
This forced us to introduce the concept of an environment, which is very
similar to the stack in normal programming languages. As we descend down
the AST, we add new identifiers bindings to this environment whenever we
encounter constructs that introduce new bindings in scope. In our case,
this is only the param of a function.