[lingua-004] Rename parser to reader
There's a small difference between a parser and a reader. The reader
reads the overall structure, looking only at parantheses and whitespace
characters. It does not attempt to classify the source code futher, for
example to create semantic objects representing a function expression,
a function call or a let definition. These are all special forms of the
language that are handled further up the chain, in the interpreter or
macro expander, not by the reader.